backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pymysql==1.0.2
python-dotenv==0.19.0
pydantic-settings==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11